add integratee recent participants contract, refactor
