+++ lib/platform-stats.js
@@ -2,6 +2,7 @@ import assert from 'node:assert'
 import createDebug from 'debug'
 import * as Sentry from '@sentry/node'
 import pRetry from 'p-retry'
+import { recentParticipantsContract } from './contracts.js'
 
 const debug = createDebug('spark:platform-stats')
 
@@ -105,6 +106,14 @@ export const updateDailyStationStats = async (
  */
 export const updateDailyParticipants = async (pgClient, participantIds) => {
   debug('Updating daily participants, count=%s', participantIds.length)
+  ;(async () => {
+    try {
+      await recentParticipantsContract.set(new Date().getDay(), participantIds)
+    } catch (err) {
+      console.error('Error updating spark-evaluations-recent-participants', err)
+      Sentry.captureException(err)
+    }
+  })()
   await pgClient.query(`
     INSERT INTO daily_participants (day, participant_id)
     SELECT now() as day, UNNEST($1::INT[]) AS participant_id
